athena/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "athena-agent"
version = "9.6.3"
description = "Athena Bionic OS — The Linux Distro for AI Agents. Model-agnostic, local-first, sovereign."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Winston Koh" }
]
keywords = [
    "ai-agent", "llm", "cognitive-os", "memory-bank",
    "model-agnostic", "ai-framework", "developer-tools"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "diskcache>=5.6.3",
    "rich>=13.0.0",
]

[project.scripts]
athena = "athena.__main__:main"

[project.urls]
Homepage = "https://winstonkoh87.com/athena/"
Repository = "https://github.com/winstonkoh87/Athena-Public"
Documentation = "https://github.com/winstonkoh87/Athena-Public#readme"
Changelog = "https://github.com/winstonkoh87/Athena-Public/blob/main/docs/CHANGELOG.md"

[project.optional-dependencies]
search = [
    "google-generativeai>=0.7.0",
    "sentence-transformers>=2.2.0",
    "flashrank>=0.2.9",
    "torch>=2.0.0",
]
cloud = [
    "supabase>=2.0.0",
]
full = [
    "athena-agent[search,cloud]",
    "google-generativeai>=0.7.0",
    "supabase>=2.0.0",
    "anthropic>=0.40.0",
    "flashrank>=0.2.9",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]
dev = [
    "pytest",
    "black",
    "ruff",
    "pipreqs",
    "pydeps",
]

[tool.hatch.build.targets.wheel]
packages = ["src/athena"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.isort]
known-first-party = ["athena"]

[tool.black]
line-length = 88
target-version = ["py310"]
